+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra.hs
@@ -7,16 +7,25 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Cardano.Ledger.Dijkstra (DijkstraEra, ApplyTxError (..)) where
+module Cardano.Ledger.Dijkstra (
+  DijkstraEra,
+  ApplyTxError (..),
+  mkDijkstraStAnnTopTx,
+) where
 
+import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusContext (mkTxInfoResult), LedgerTxInfo (..))
+import Cardano.Ledger.Alonzo.Plutus.Evaluate (collectPlutusScriptsWithContext)
+import Cardano.Ledger.Alonzo.UTxO (AlonzoEraUTxO, AlonzoScriptsNeeded)
 import Cardano.Ledger.BaseTypes (Inject (inject))
 import Cardano.Ledger.Binary (DecCBOR, EncCBOR)
 import Cardano.Ledger.Block (EraBlockHeader)
 import Cardano.Ledger.Conway.Governance (RunConwayRatify)
 import Cardano.Ledger.Dijkstra.BlockBody ()
+import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Era
 import Cardano.Ledger.Dijkstra.Forecast ()
 import Cardano.Ledger.Dijkstra.Genesis ()
@@ -30,7 +39,7 @@ import Cardano.Ledger.Dijkstra.State.CertState ()
 import Cardano.Ledger.Dijkstra.State.Stake ()
 import Cardano.Ledger.Dijkstra.Transition ()
 import Cardano.Ledger.Dijkstra.Translation ()
-import Cardano.Ledger.Dijkstra.Tx ()
+import Cardano.Ledger.Dijkstra.Tx (DijkstraStAnnTx (..))
 import Cardano.Ledger.Dijkstra.TxBody ()
 import Cardano.Ledger.Dijkstra.TxInfo ()
 import Cardano.Ledger.Dijkstra.TxWits ()
@@ -41,14 +50,23 @@ import Cardano.Ledger.Shelley.API (
   ApplyTx (..),
   ruleApplyTxValidation,
  )
+import Cardano.Ledger.State (EraUTxO (..), ScriptsProvided, UTxO)
+import Cardano.Slotting.EpochInfo (EpochInfo)
+import Cardano.Slotting.Time (SystemStart)
 import Data.Bifunctor (Bifunctor (first))
+import Data.Foldable (toList)
 import Data.List.NonEmpty (NonEmpty)
+import Data.Text (Text)
 import GHC.Generics (Generic)
+import Lens.Micro
 
 instance ApplyTx DijkstraEra where
   newtype ApplyTxError DijkstraEra = DijkstraApplyTxError (NonEmpty (DijkstraMempoolPredFailure DijkstraEra))
     deriving (Eq, Show)
     deriving newtype (EncCBOR, DecCBOR, Semigroup, Generic)
+
+  mkStAnnTx = mkDijkstraStAnnTopTx
+
   applyTxValidation validationPolicy globals env state tx =
     first DijkstraApplyTxError $
       ruleApplyTxValidation @"MEMPOOL" validationPolicy globals env state tx
@@ -67,3 +85,68 @@ instance Inject (NonEmpty (DijkstraMempoolPredFailure DijkstraEra)) (ApplyTxErro
 
 instance Inject (NonEmpty (DijkstraLedgerPredFailure DijkstraEra)) (ApplyTxError DijkstraEra) where
   inject = DijkstraApplyTxError . fmap LedgerFailure
+
+mkDijkstraStAnnTopTx ::
+  ( AlonzoEraUTxO era
+  , AlonzoEraTx era
+  , DijkstraEraTxBody era
+  , EraPlutusContext era
+  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
+  ) =>
+  EpochInfo (Either Text) ->
+  SystemStart ->
+  PParams era ->
+  UTxO era ->
+  Tx TopTx era ->
+  DijkstraStAnnTx TopTx era
+mkDijkstraStAnnTopTx ei sysStart pp utxo tx =
+  let
+    txBody = tx ^. bodyTxL
+    scriptsNeeded = getScriptsNeeded utxo txBody
+    scriptsProvided = getScriptsProvided utxo tx
+    stAnnSubTxs =
+      map
+        (mkDijkstraStAnnSubTx ei sysStart pp utxo scriptsProvided)
+        (toList (txBody ^. subTransactionsTxBodyL))
+   in
+    DijkstraStAnnTopTx
+      { dsattTx = tx
+      , dsattProtocolVersion = pp ^. ppProtocolVersionL
+      , dsattScriptsNeeded = scriptsNeeded
+      , dsattScriptsProvided = scriptsProvided
+      , dsattPlutusScriptsWithContext = collectPlutusScriptsWithContext ei sysStart pp tx utxo
+      , dsattStAnnSubTxs = stAnnSubTxs
+      }
+
+mkDijkstraStAnnSubTx ::
+  ( AlonzoEraUTxO era
+  , AlonzoEraTx era
+  , EraPlutusContext era
+  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
+  ) =>
+  EpochInfo (Either Text) ->
+  SystemStart ->
+  PParams era ->
+  UTxO era ->
+  ScriptsProvided era ->
+  Tx SubTx era ->
+  DijkstraStAnnTx SubTx era
+mkDijkstraStAnnSubTx ei sysStart pp utxo scriptsProvided tx =
+  let
+    scriptsNeeded = getScriptsNeeded utxo (tx ^. bodyTxL)
+    ledgerTxInfo =
+      LedgerTxInfo
+        { ltiProtVer = pp ^. ppProtocolVersionL
+        , ltiEpochInfo = ei
+        , ltiSystemStart = sysStart
+        , ltiUTxO = utxo
+        , ltiTx = tx
+        }
+   in
+    DijkstraStAnnSubTx
+      { dsastTx = tx
+      , dsastScriptsNeeded = scriptsNeeded
+      , dsastScriptsProvided = scriptsProvided
+      , dsastTxInfoResult = mkTxInfoResult ledgerTxInfo
+      , dsastPlutusScriptsWithContext = collectPlutusScriptsWithContext ei sysStart pp tx utxo
+      }