+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.3.0.0
 
+* Add `DijkstraStAnnTx`
+* Add `scriptsProvidedDijkstraStAnnTx`
+* Add `mkDijkstraStAnnTopTx`
 * Remove `ToCBOR` and `FromCBOR` instances for `DijkstraGovPredFailure`
 * Add `getDijkstraScriptsProvided`
 * Add `MissingRequiredGuards` constructor to `DijkstraUtxowPredFailure`